Change view/model to use DeferredGameAction
References within a view are often lost on orientation change. Items
disconnected from a model i.e. potential actions to take are especially
bitten by this. To start solving this, defer the inclusion of the
(disconnectable) model in view actions until the last possible moment.
For now, these are reconnected in the ApplicationView; this should
probably be changed later.