Contracted Intent/Bundle interface to a single Wrapper class
It turns out that Android limits Intent/Bundle sizes heavily. This is
unfortunate when I am attempting to shuttle around large models (*cough*
inventory *cough*). We will have to store unprocessed models in local
memory instead, which is kind of a hack. For now, all read/writes of
models to intents have been contained to a single class.